test(duckdb): Move DuckDB integration test (#737)
