+++ internal/impl/sql/integration_test.go
@@ -1683,52 +1683,3 @@ output:
 		return countAfterRestart > count+100
 	}, time.Minute, time.Second)
 }
-
-func TestIntegrationDuckDBAppender(t *testing.T) {
-	integration.CheckSkip(t)
-
-	db := filepath.Join(t.TempDir(), "vault.duckdb")
-
-	streamBuilder := service.NewStreamBuilder()
-	require.NoError(t, streamBuilder.SetYAML(fmt.Sprintf(`
-input:
-  generate:
-    mapping: |
-      root.duck = ["Scrooge McDuck", "Donald Duck", "Daffy Duck"][counter() %% 3]
-      root.gold_coins = counter() * 10
-    count: 5
-    interval: ""
-
-output:
-  duckdb_append:
-    dsn: %s
-    table: vault_deposits
-    columns: [duck, gold_coins]
-    args_mapping: "root = [this.duck, this.gold_coins]"
-    init_statement: |
-      CREATE TABLE IF NOT EXISTS vault_deposits (
-        duck       VARCHAR,
-        gold_coins BIGINT
-      )
-
-logger:
-  level: none
-`, db)))
-
-	stream, err := streamBuilder.Build()
-	require.NoError(t, err)
-	require.NoError(t, stream.Run(context.Background()))
-
-	// open a fresh connection to verify the written data
-	dbConn, err := sql.Open("duckdb", db)
-	require.NoError(t, err)
-	defer dbConn.Close()
-
-	var totalDeposits int
-	require.NoError(t, dbConn.QueryRow(`SELECT COUNT(*) FROM vault_deposits`).Scan(&totalDeposits))
-	assert.Equal(t, 5, totalDeposits)
-
-	var totalCoins int64
-	require.NoError(t, dbConn.QueryRow(`SELECT SUM(gold_coins) FROM vault_deposits`).Scan(&totalCoins))
-	assert.EqualValues(t, 10+20+30+40+50, totalCoins)
-}
